src/requirements.txt
chardet==2.3.0
Django==2.0.4
django-annoying==0.10.4
mysqlclient==1.3.10
Pillow==5.1.0
pycryptodome==3.6.1
pycurl==7.43.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.9.1
six==1.10.0
sorl-thumbnail==12.4.1
urllib3==1.13.1